+++ app/views/layouts/application.html.erb
@@ -45,12 +45,12 @@
                 </li>
                 <% if user_signed_in? %>
                   <li>
-                    <%= link_to(home_index_path) do %>
+                    <%= link_to(posts_path) do %>
                       <i class="fas fa-book-open"></i>
                     <% end %>
                   </li>
                   <li>
-                    <%= link_to(home_new_path) do %>
+                    <%= link_to(new_post_path) do %>
                       <i class="fas fa-pen-nib"></i>
                     <% end %>
                   </li>
