+++ app/views/posts/edit.html.erb
@@ -1,7 +1,7 @@
 <div class="container-fluid">
     <div class="row">
         <div class="center-block mx-auto">
-            <%= form_tag("/home/#{@post.id}/update",{method: :post}) do %>
+            <%= form_tag("/posts/#{@post.id}",{method: :patch}) do %>
                 <%= render 'shared/post_form' %>
             <% end %>
         </div>
