+++ app/views/posts/index.html.erb
