+++ app/views/posts/new.html.erb
@@ -1,7 +1,7 @@
 <div class="container-fluid">
   <div class="row">
         <div class="center-block mx-auto">
-                <%= form_tag(home_create_path) do %>
+                <%= form_tag(posts_path,{method: :post}) do %>
                     <%= render 'shared/post_form' %>
                 <% end %>
         </div>
